Create tacoCat
You run an insanely profitable business making gourmet tacos for cats. Given an object that represents the inventory of your company, add a method 'currentInventory' to your object that returns the current value of your company's inventory.

Then, add a sale method that takes an order object, returns the price of the order, updates the inventory of the purchased items, and adds the price of the order to the company's cash property.